CV_final/requirements.txt
opencv-python
numpy
matplotlib
pillow
transformers
torch
tqdm
scipy
scikit-image

